requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.7.4
aiodns>=3.0.0
pydantic>=1.8.0
python-dotenv>=0.19.0